pm/drivers/clock: fix MISRA.ETYPE.COMP.ASSIGN.2012
The value of a composite expression shall not be
assigned to an object with wider essential type.

Fixed by typecasting to the required datatype.

Fixes: SYSFW-6414

+++ pm/drivers/clock/clk_pll_16fft.c
@@ -850,7 +850,7 @@ static u32 clk_pll_16fft_internal_set_freq_from_pll_table(struct clk *pll_clk,
 			 *
 			 * Note: We break up the calculation in order to avoid a div64.
 			 */
-			clkod_plld = soc_pll_table[data_pll->pll_entries[i]].plld * soc_pll_table[data_pll->pll_entries[i]].clkod;
+			clkod_plld = (u64) (soc_pll_table[data_pll->pll_entries[i]].plld * soc_pll_table[data_pll->pll_entries[i]].clkod);
 
 			actual64 = ((u64) (input / clkod_plld)) * soc_pll_table[data_pll->pll_entries[i]].pllm;
 			rem64 = ((u64) (input % clkod_plld)) * soc_pll_table[data_pll->pll_entries[i]].pllm;
